# docs/api_references/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meeting Minutes Agent API",
    "description": "会议纪要 Agent - 自动生成会议纪要的 AI 服务",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/prompt-templates": {
      "get": {
        "tags": [
          "prompt-templates"
        ],
        "summary": "List Prompt Templates",
        "description": "列出所有可用的提示词模板\n\n- 支持按作用域过滤 (global/private)\n- 支持按内容类型过滤\n- 全局模板对所有用户可见\n- 私有模板只对创建者可见",
        "operationId": "list_prompt_templates_api_v1_prompt_templates_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "作用域过滤 (global/private)",
              "title": "Scope"
            },
            "description": "作用域过滤 (global/private)"
          },
          {
            "name": "artifact_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "内容类型过滤",
              "title": "Artifact Type"
            },
            "description": "内容类型过滤"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "用户 ID (用于查询私有模板)",
              "title": "User Id"
            },
            "description": "用户 ID (用于查询私有模板)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPromptTemplatesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "prompt-templates"
        ],
        "summary": "Create Prompt Template",
        "description": "创建私有提示词模板\n\n- 只能创建私有模板 (scope=private)\n- 模板 ID 自动生成\n- 绑定到创建者的 user_id",
        "operationId": "create_prompt_template_api_v1_prompt_templates_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "用户 ID (创建者)",
              "title": "User Id"
            },
            "description": "用户 ID (创建者)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePromptTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prompt-templates/{template_id}": {
      "get": {
        "tags": [
          "prompt-templates"
        ],
        "summary": "Get Prompt Template",
        "description": "获取提示词模板详情\n\n- 全局模板对所有用户可见\n- 私有模板只对创建者可见",
        "operationId": "get_prompt_template_api_v1_prompt_templates__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "用户 ID (用于验证私有模板权限)",
              "title": "User Id"
            },
            "description": "用户 ID (用于验证私有模板权限)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptTemplateDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "prompt-templates"
        ],
        "summary": "Update Prompt Template",
        "description": "更新私有提示词模板\n\n- 只能更新自己创建的私有模板\n- 不能更新全局模板",
        "operationId": "update_prompt_template_api_v1_prompt_templates__template_id__put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "用户 ID (用于验证权限)",
              "title": "User Id"
            },
            "description": "用户 ID (用于验证权限)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePromptTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "prompt-templates"
        ],
        "summary": "Delete Prompt Template",
        "description": "删除私有提示词模板\n\n- 只能删除自己创建的私有模板\n- 不能删除全局模板",
        "operationId": "delete_prompt_template_api_v1_prompt_templates__template_id__delete",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "用户 ID (用于验证权限)",
              "title": "User Id"
            },
            "description": "用户 ID (用于验证权限)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePromptTemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreatePromptTemplateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "模板标题"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "模板描述"
          },
          "prompt_body": {
            "type": "string",
            "title": "Prompt Body",
            "description": "提示词正文(含占位符)"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type",
            "description": "生成内容类型"
          },
          "supported_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Languages",
            "description": "支持的语言列表",
            "default": [
              "zh-CN"
            ]
          },
          "parameter_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameter Schema",
            "description": "参数定义"
          }
        },
        "type": "object",
        "required": [
          "title",
          "prompt_body",
          "artifact_type",
          "parameter_schema"
        ],
        "title": "CreatePromptTemplateRequest",
        "description": "创建提示词模板请求",
        "example": {
          "artifact_type": "meeting_minutes",
          "description": "适用于技术团队的会议纪要",
          "parameter_schema": {
            "meeting_description": {
              "default": "",
              "description": "会议描述信息",
              "required": false,
              "type": "string"
            }
          },
          "prompt_body": "你是一个专业的会议纪要助手。\n\n会议信息:\n{meeting_description}\n\n请生成技术导向的会议纪要...",
          "supported_languages": [
            "zh-CN",
            "en-US"
          ],
          "title": "我的自定义会议纪要模板"
        }
      },
      "CreatePromptTemplateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "提示词模板已创建"
          }
        },
        "type": "object",
        "required": [
          "success",
          "template_id"
        ],
        "title": "CreatePromptTemplateResponse",
        "description": "创建提示词模板响应"
      },
...
      "DeletePromptTemplateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "提示词模板已删除"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "DeletePromptTemplateResponse",
        "description": "删除提示词模板响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListPromptTemplatesResponse": {
        "properties": {
          "templates": {
            "items": {
              "$ref": "#/components/schemas/PromptTemplate"
            },
            "type": "array",
            "title": "Templates",
            "description": "提示词模板列表"
          }
        },
        "type": "object",
        "required": [
          "templates"
        ],
        "title": "ListPromptTemplatesResponse",
        "description": "列出提示词模板响应"
      },
...
      "PromptTemplate": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "description": "模板唯一标识"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "模板标题"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "用户可读说明"
          },
          "prompt_body": {
            "type": "string",
            "title": "Prompt Body",
            "description": "提示词正文(含占位符,如 {meeting_description})"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type",
            "description": "生成内容类型(meeting_minutes/action_items/summary_notes)"
          },
          "supported_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Languages",
            "description": "支持的语言列表",
            "default": [
              "zh-CN",
              "en-US"
            ]
          },
          "parameter_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameter Schema",
            "description": "参数定义(参数名 -> {type, required, default})"
          },
          "is_system": {
            "type": "boolean",
            "title": "Is System",
            "description": "是否为系统预置模板",
            "default": true
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "description": "作用域(global/private)",
            "default": "global"
          },
          "scope_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Id",
            "description": "作用域 ID(user_id)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "title",
          "description",
          "prompt_body",
          "artifact_type",
          "parameter_schema"
        ],
        "title": "PromptTemplate",
        "description": "提示词模板",
        "example": {
          "artifact_type": "meeting_minutes",
          "description": "生成包含摘要、关键要点和行动项的标准会议纪要",
          "is_system": true,
          "parameter_schema": {
            "meeting_description": {
              "default": "",
              "description": "会议描述信息",
              "required": false,
              "type": "string"
            }
          },
          "prompt_body": "你是一个专业的会议纪要助手。\n\n会议信息:\n{meeting_description}\n\n请根据以下会议转写生成纪要...",
          "supported_languages": [
            "zh-CN",
            "en-US"
          ],
          "template_id": "tpl_001",
          "title": "标准会议纪要"
        }
      },
      "PromptTemplateDetailResponse": {
        "properties": {
          "template": {
            "$ref": "#/components/schemas/PromptTemplate"
          }
        },
        "type": "object",
        "required": [
          "template"
        ],
        "title": "PromptTemplateDetailResponse",
        "description": "提示词模板详情响应"
      },
...
      "UpdatePromptTemplateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "模板标题"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "模板描述"
          },
          "prompt_body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Body",
            "description": "提示词正文"
          },
          "supported_languages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supported Languages",
            "description": "支持的语言列表"
          },
          "parameter_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameter Schema",
            "description": "参数定义"
          }
        },
        "type": "object",
        "title": "UpdatePromptTemplateRequest",
        "description": "更新提示词模板请求"
      },
      "UpdatePromptTemplateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "提示词模板已更新"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "UpdatePromptTemplateResponse",
        "description": "更新提示词模板响应"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}